Example for remapping a dataflow
The goal of this script is to illustrate FMR API capabilities (not designed to be used in production but meant to be usable by anyone for getting a better understanding on the API).  The idea would be to create a new dataflow with a constraint to slice the original EXR1 DSD and getting only Annual data (can be done by querying partial data from the original source but here, it is just an example for using the API).

The FMR docker version comes by default with two dataflows for illustration.
They are not fully identical  to the ones on ECB website.
So this script replaces the original content by the content available on ECB website.

Unfortunately, EXR has an issue on the frequency (ECB uses H code for semi-annual data). So, the idea is to take ownership with a different agency (I used SDMX agency for simplicity for should be your own agency). And to use a mapping (all mappings are implicit except the FREQ dimension where a mapping representation has been defined).

First result: The GUI doesn't seem to allow a conversion from one agency to another one.
Second result: Using the API, the revalidate endpoint doesn't seem to correct the issue on the frequency.